requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.3.1
graphviz==0.15
gunicorn==20.0.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.2.1
mysqlclient==2.0.2
nltk==3.5
numpy==1.18.4
opencv-python==4.4.0.46
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pydotplus==2.0.2
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.48.2
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.2.0